% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_N_map}
\alias{plot_N_map}
\title{plot a map of estimated abundance or related quantity.}
\usage{
plot_N_map(N, Grid, highlight = NULL, leg.title = "Abundance")
}
\arguments{
\item{N}{quantity to plot}

\item{Grid}{SpatialPolygonsDataFrame holding locations of each grid cell}

\item{highlight}{[optional] A vector of cells to highlight.  Default = NULL,}

\item{leg.title}{Title of the legend}
}
\value{
adjacency matrix
}
\description{
plot a map of estimated abundance or related quantity.
}
\author{
Paul Conn \email{paul.conn@noaa.gov}
}
\keyword{adjacency}

