% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim_density_PB}
\alias{sim_density_PB}
\title{simulate density for seals and polar bears for Chukchi power analysis}
\usage{
sim_density_PB(Grid_chukchi)
}
\arguments{
\item{Grid_chukchi}{A list object consisting of "Grid" - a SpatialPolygonsDataFrame with housing
different sampling units (grid cells), and "Adj" - an adjacency matrix describing neighborhood structure}
}
\value{
a list giving expected abundance (Lambda.true) and simulated random effects (Eta.true)
}
\description{
simulate density for seals and polar bears for Chukchi power analysis
}
\author{
Paul B. Conn
}
\keyword{abundance}
\keyword{simulation}

