\docType{data}
\name{Sim_effort}
\alias{Sim_effort}
\title{Aerial survey effort for different hypothetical flight combinations for Chukchi Sea power analysis}
\description{
  A list object, with each element of the list housing the survey units   and associated areas surveyed in each for a particular flight and effort allocation strategy. Each element of the list is a data.frame with two columns, consisting of "poly" (which survey unit/grid cell is sampled), and "area" - the proportional area in each grid cell that is sampled.
}
\author{
  Paul Conn \email{paul.conn@noaa.gov}
}
\keyword{data}
\keyword{datasets}

