% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulate_counts_PB}
\alias{simulate_counts_PB}
\title{given underlying density, effort, and observation process information, simulate animal counts}
\usage{
simulate_counts_PB(Lambda, Effort, Thin)
}
\arguments{
\item{Lambda}{A (3 x n.cells) matrix holding expected abundance for bearded seals, ringed seals, and polar bear in each grid cell}

\item{Effort}{A matrix or data.frame giving effort in sampled cells.  Columns are "poly", giving which cells are sampled, and "area", the proportion of each sampled grid cell that is surveyed}

\item{Thin}{A vector giving an additional thinning factor for each species (corresponding to detection and availability)}
}
\value{
A data frame holding animal observation data.  Columns include
    -"Transect" (which cell the observation occurred in)
    -"Photo" (whether a photo was obtained - here these are all set to 1.0)
    -"Obs" (what species was observed), and
    -"Group" (group size for each hotspot) - here assumed to by 1.0
}
\description{
given underlying density, effort, and observation process information, simulate animal counts
}
\author{
Paul B. Conn
}
\keyword{count}
\keyword{data}
\keyword{simulation,}

