\docType{data}
\name{Thin}
\alias{Thin}
\title{Posterior distributions for thinning parameters}
\description{
  Each row includes a posterior distribution for the thinning parameter
  for each species (row 1: bearded, row 2: ringed, row 3: polar bear).
  This is a product of posterior distributions for detection and availability (haulout   probability)
}
\author{
  Paul Conn \email{paul.conn@noaa.gov}
}
\keyword{data}
\keyword{datasets}

